Update Rust to 1.66.0
Rust has been updated to its latest version, and a few new Clippy
warnings have been fixed in the process.

+++ game/src/systems/setup_grid.rs
@@ -13,11 +13,7 @@ pub fn setup_grid(map: Res<Map>, mut commands: Commands) {
 
         commands.spawn_bundle(SpriteBundle {
             transform: Transform {
-                translation: Vec3::new(
-                    point.x() as f32,
-                    point.y() as f32,
-                    RenderLayer::RoutingGrid.z(),
-                ),
+                translation: Vec3::new(point.x(), point.y(), RenderLayer::RoutingGrid.z()),
                 scale: Vec3::new(2.0, 2.0, 0.0),
                 ..Default::default()
             },
